move go source from src/main/go to src/go
