+++ Android/app/src/go/intra/doh/atomic.go
