+++ Android/app/src/go/intra/split/direct_split.go
