+++ Android/app/src/go/intra/split/example/main.go
@@ -22,7 +22,7 @@ import (
 	"net"
 	"os"
 
-	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/split"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/go/intra/split"
 )
 
 func main() {
