+++ Android/app/src/go/intra/stream_dialer.go
@@ -23,9 +23,9 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/doh"
-	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/protect"
-	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/split"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/go/intra/doh"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/go/intra/protect"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/go/intra/split"
 	"github.com/Jigsaw-Code/outline-sdk/transport"
 )
